refactor(HoverObject): Move HoverObject to its own folder
